grus: extract-files: Properly patch libdpmframework.so
Replacing ELF dependency libhidltransport.so by libcutils-v29.so
to resolve missing 'strdup8to16' symbol is not really good.

The mentioned symbol was removed in API level 30 so the proper way
is adding a dependency on libcutils-v29.so by patchelf. But we need
to preserve libcutils.so as well because of new symbols.

The method is tricky because SONAME of libcutils-v29.so is
also 'libcutils.so' so we need to replace the dependency on
libcutils.so by libcutils-v29.so first and then add again
dependency on libcutils.so so the real v30 libcutils.so will
be loaded first.

Signed-off-by: Ivan Vecera <ivan@cera.cz>
Change-Id: I62cb7c5a0c04333725e09a6376afd09694923245