grus: audio: Use reference for all default volume table
This patch migrates all default volume tables as reference
volume tables.
No implicit volume tables is allowed any more, the reference
must be given for all stream type and device category.
It allows removing the WA that was introduce to use the stream CNT
to host the default volume tables.

Change-Id: I19f5e9d2a62d448f0f8447f93acbbd2e1a723958
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Signed-off-by: Alex Damaratski <alexeydomoratsky1@gmail.com>